重构: ScriptEditor.vue 拆分为 9 个子组件，ChapterFlow 伪元素迁移至 Tailwind
